Compress images when adding them to the archive
The compression step uses Google's guetzli, which is very effective but
also very slow.  It also isn't readily available in binary form, though the
source is easy to compile, especially on Linux.  The archiver script
expects to find guetzli in the PATH.

Currently, it uses a quality of 95 for all images.

It aggressively uses Python's multiprocessing library to parallelize the
sluggish compression work.  We'll see how well it works, I suppose.  It can
be turned off by replacing the main_parallel call with the main_single
call at the bottom of generate_pages.py.