Add a "superoptimizing" process in the image compressor
If we're resampling *AND* compressing, then sometimes the result image is
larger than the original image, but if we instead compress the original
non-resampled file, we get a smaller file.  We need to find the smaller
file, especially because there may be other places referencing the same
image without resampling.  We were ending up with situations like this:

img-100x200-g95.jpg   10000
img-g95.jpg           20000
img.jpg               30000

An img tag sized at 200x400 might have referred to img.jpg, but an a href
linking to the image would link to img-g95.jpg.  Instead, the img tag
should refer to img-g95.jpg, and img.jpg can be left out of the archive.

I don't like how this code has turned out, but I'm not immediately sure
how to fix it.  I *think* it's correct.  It's possible for multiple jobs
to overwrite the same files, and it's even possible for multiple guetzli
jobs to run with the same input and output.  (The jobs will have different
temporary output files, and the renaming at the end happens under
_fs_lock.)