Don't convert GIFs to PNG if we don't need to do anything with them
The GIF->PNG conversion is actually breaking with the "chatbox" icon that
appears to the left of each comment.  The conversion adds a vertical line
to the right column of the transparent image.

https://resources.blogblog.com/blogblog/data/rounders4/icon_comment_left.gif

In principle, this isn't the right fix, but it's better behavior that
happens to fix the problem too.