Include a post's title in its comment permalinks
e.g.

Instead of:

   #c100
   #c101

Use:

   the-post-name.html#c100
   the-post-name.html#c101

The problem is that the existing Blogger permalinks have the form:

   https://site.com/the-post-name.html?showComment=4000#c100
   https://site.com/the-post-name.html?showComment=4001#c101

If a visitor follows the Blogger permalink to the archive site, then
tries to make another permalink, their browser will combine the old
showComment parameter with the new anchor:

   https://site.com/the-post-name.html?showComment=4000#c101
   https://site.com/the-post-name.html?showComment=4001#c100

AFAIK, it doens't matter, because I can't get the showComment parameter to
*do* anything on the Blogger site.  It seems that only the anchor has an
effect.  Still, to reduce confusion, the parameter should be omitted from
the archive's permalinks.